refactor joystick serial
