+++ src/components/common/Camera.tsx
