+++ src/components/operational-view-alpha/OVA-OverlayMenu.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-import Camera from "../Camera";
-import Status from "../Status";
+import Camera from "../common/Camera";
+import Status from "../common/Status";
 import { useState } from "react";
-import "./OV1-MainView";
+import "./OVA-MainView";
 function OverlayMenu() {
   return (
     <div className="OverlayMenu">
