+++ src/components/operational-view-alpha/OperationalViewAlpha.css
@@ -1,7 +1,7 @@
-body {
+/* body {
   margin: 5px;
   background: rgb(130, 163, 157);
- }
+ } */
 
  .OperationalView{
   width: 100vw;
