+++ src/components/prototype-assets/Wifi.tsx
