+++ src/index.tsx
@@ -2,8 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
 import './index.css';
-import './components/OperationalView1.css' 
-
+import './components/operational-view-alpha/OperationalViewAlpha'
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
 root.render(
   <React.StrictMode>
